+++ author.php
@@ -1,5 +1,11 @@
-<?php while (have_posts()) : the_post(); ?>
-  <?php get_template_part('templates/content-author', get_post_type() != 'post' ? get_post_type() : get_post_format()); ?>
-<?php endwhile; ?>
+<?php
+while (have_posts()) {
+	the_post();
+	get_template_part(
+		'templates/content-author',
+		 get_post_type() != 'post' ? get_post_type() : get_post_format()
+	 );
+};
 
-<?php the_posts_navigation(); ?>
+the_posts_navigation();
+?>